modified memory index to support long doc ids, more work is still required, but can move on to changing the iterator tree to match, and bugs will be found

+++ core/src/main/java/org/lemurproject/galago/core/index/mem/MemoryPositionalIndex.java
@@ -498,7 +498,7 @@ public class MemoryPositionalIndex implements MemoryIndexPart, AggregateIndexPar
         currDocument = postings.lastDocument;
         currCount = postings.lastCount;
       } else {
-        currDocument += documents_reader.readInt();
+        currDocument += documents_reader.readLong();
         currCount = counts_reader.readInt();
       }
       loadExtents();
@@ -686,7 +686,7 @@ public class MemoryPositionalIndex implements MemoryIndexPart, AggregateIndexPar
         currDocument = postings.lastDocument;
         currCount = postings.lastCount;
       } else {
-        currDocument += documents_reader.readInt();
+        currDocument += documents_reader.readLong();
         currCount = counts_reader.readInt();
       }
