+++ core/src/main/java/org/lemurproject/galago/core/index/mem/MemoryWindowIndex.java
@@ -483,7 +483,7 @@ public class MemoryWindowIndex implements MemoryIndexPart, AggregateIndexPart {
         currDocument = postings.lastDocument;
         currCount = postings.lastCount;
       } else {
-        currDocument += documents_reader.readInt();
+        currDocument += documents_reader.readLong();
         currCount = counts_reader.readInt();
       }
       loadExtents();
